Set of fixes after steaming test
- Add temprerature sensor back to 002
- Make both 002 and 1204 temperature sensors for now, todo - clean up and combine
- Change RCCState, reduce some fileds to 8bit, incread distance and tick to 32b
- Update Station code to deal with new firmat of RCCState
- Revert speed sensor to use floats instead of doubles
- Flash 'neutral' on the Throttle screen (so it is clear why loco doesn't move)
- Add XX.Xk for distance on Throttle screen
- Clen up 'big font' screen
- Reset timer by key '3'
- Change screen saver to 3 mins (from 1)

+++ RRLibs/SpeedSensor.h
@@ -58,30 +58,30 @@ class SpeedSensorLegacy: public SpeedSensorBase {
   private:
     const long MAX_PERIOD = 3000;
     const int DEBOUNCE_THRESHOLD = 20;
-    const double SCALE = 2;
+    const float SCALE = 2;
     SimpleKalmanFilter filter;
     int pin;
 
     volatile long period;
     volatile long lastTimestamp;
 
-    double speed;
-    double rotationPerSecond;
-    long distance;
+    float speed;
+    float rotationPerSecond;
+    float distance;
 
   public:
-    double measurementError;
+    float measurementError;
     SpeedSensorLegacy(int pin) : pin(pin), filter(5, 5, 0.5) {
     }
     void setup() {
       pinMode(pin, INPUT_PULLUP);
       attachInterrupt(digitalPinToInterrupt(pin), speedHandler, FALLING);
     }
-    void setKalmanError(double value) {
+    void setKalmanError(float value) {
       measurementError = value;
       filter.setMeasurementError(value);
     }
-    void setKalmanNoise(double value) {
+    void setKalmanNoise(float value) {
       filter.setProcessNoise(value);
     }
     float getSpeedRaw() {
